Checked tags cache, optional hardlinking for duplicates
If you have lots of folders to download, and have common tags
for many folders, each common tag was checked again for new folders.
Now checked tags are cached.

If you download many folders with many duplicates, they can take
a lot of space. If we make hard links instead of file copies,
we can save the space, but if file corrupts, no backup for you.
So 'make_hardlinks' option was added to config.ini in '[Other]' section,
so you can have a choice between space or redundancy.